Improve engine responsibilities
- Changed the central point to be the engine instead of window
this change was necessary to garantee that the engine always have the
control of the app main flow.

- Changed the physics engine to be more controllable by the engine itself
now we can control the amount of ticks that are upated by the engine to increase
or decrease the physics precision. Remember that the number of steps will decrease the engine performance if it's high